Resize images, and add some color to links
I resized the images to make it more clean, and not messy taking up the
whole screen. Added some colors to the recipe links also that match with
the foods that are in the links. Changed the fonts to Times New Roman,
Times, and serif. Did all this while learning the basics of CSS.